--- a/app/bus/search/context.ts
+++ b/app/bus/search/context.ts
@@ -1,7 +1,0 @@
-// Core
-import { createContext } from 'react';
-
-// Types
-import type { ReturnHook } from './hooks/types';
-
-export default createContext<Partial<ReturnHook>>([]);