--- a/app/bus/search/hooks/types.ts
+++ b/app/bus/search/hooks/types.ts
@@ -1,4 +1,0 @@
-// Types
-import { MouseEvent } from 'react'
-
-export type ReturnHook = [string, (event: MouseEvent<HTMLInputElement>) => void]