--- a/app/components/search/index.tsx
+++ b/app/components/search/index.tsx
@@ -1,14 +1,22 @@
 // Core
-import { FC } from 'react';
+import { FC, useCallback } from 'react';
 
 // Components
 import Input from '../input';
 
-// Bus
-import useContextSearch from '~/bus/search/hooks/use-context-search';
+// Context
+import useSearchState from '~/context/search/hooks/use-search-state';
+import useSearchActions from '~/context/search/hooks/use-search-actions';
 
 const Search: FC = () => {
-  const [search, onChangeSearch] = useContextSearch();
+  const search = useSearchState();
+  const setSearch = useSearchActions();
+
+  const onChangeSearch = useCallback((event) => {
+    if (setSearch) {
+      setSearch(event.target.value);
+    }
+  }, [setSearch]);
 
   return <Input placeholder="Search..." onChange={onChangeSearch} value={search} />
 }