--- a/app/context/search/context.tsx
+++ b/app/context/search/context.tsx
@@ -1,0 +1,12 @@
+// Core
+import { createContext } from 'react';
+
+// Types
+import { Context } from './types';
+
+const initContext: Context = {
+  search: '',
+  setSearch: () => { }
+}
+
+export default createContext<Context>(initContext);