--- a/app/context/search/hooks/use-search-actions.ts
+++ b/app/context/search/hooks/use-search-actions.ts
@@ -1,0 +1,13 @@
+// Core
+import { useContext } from 'react';
+
+// Context
+import SearchContext from '../context';
+
+const useSearchActions = () => {
+  const { setSearch } = useContext(SearchContext);
+
+  return setSearch;
+}
+
+export default useSearchActions;