--- a/app/context/search/hooks/use-search-state.ts
+++ b/app/context/search/hooks/use-search-state.ts
@@ -1,0 +1,13 @@
+// Core
+import { useContext } from 'react';
+
+// Context
+import SearchContext from '../context';
+
+const useSearchState = () => {
+  const { search } = useContext(SearchContext);
+
+  return search;
+}
+
+export default useSearchState;