--- a/app/context/search/types.ts
+++ b/app/context/search/types.ts
@@ -1,0 +1,11 @@
+// Core
+import { ReactNode } from 'react';
+
+export interface PropsProvider {
+  children: ReactNode
+}
+
+export interface Context {
+  search: string;
+  setSearch: Function;
+}