Search optimization.

old: `12 ms - 2.3 ms - 1.6 ms - 8.1 ms`.
new: `9 ms - 1.5 ms - 0.8 ms - 0.9 ms`.